test/support/builders/metadata: Extract dir/file common build logic
No duplicate logic anymore.
And adding common behavior will be easier (e.g. detecting platform
incompatibilities automatically).
