+++ test/support/builders/metadata/file.js
@@ -1,21 +1,17 @@
 /* @flow */
 
 const crypto = require('crypto')
-const _ = require('lodash')
 
 const BaseMetadataBuilder = require('./base')
-const { assignId } = require('../../../../core/metadata')
-
-const pouchdbBuilders = require('../pouchdb')
 
 /*::
 import type Pouch from '../../../../core/pouch'
-import type { Metadata } from '../../../../core/metadata'
 */
 
 module.exports = class FileMetadataBuilder extends BaseMetadataBuilder {
   /*::
   fileOpts: {
+    docType: 'file',
     size: number,
     md5sum: string
   }
@@ -24,6 +20,7 @@ module.exports = class FileMetadataBuilder extends BaseMetadataBuilder {
   constructor (pouch /*: ?Pouch */) {
     super(pouch)
     this.fileOpts = {
+      docType: 'file',
       size: 0,
       md5sum: '1B2M2Y8AsgTpgAmY7PhCfg==' // empty
     }
@@ -36,18 +33,7 @@ module.exports = class FileMetadataBuilder extends BaseMetadataBuilder {
     return this
   }
 
-  build () /*: Metadata */ {
-    const doc = _.merge({
-      _id: '',
-      // _rev: pouchdbBuilders.rev(),
-      docType: 'file',
-      remote: {
-        _id: pouchdbBuilders.id(),
-        _rev: pouchdbBuilders.rev()
-      },
-      tags: []
-    }, this.opts, this.fileOpts)
-    assignId(doc)
-    return doc
+  attributesByType () /*: * */ {
+    return this.fileOpts
   }
 }
